@@ -7,8 +7,7 @@
   "bin": "./bin/cli.js",
   "files": [
     "bin/*",
-    "lib/*",
-    "plugin/*"
+    "dist/*"
   ],
   "repository": {
     "type": "git",
@@ -37,6 +36,16 @@
   "devDependencies": {
     "@rollup/plugin-commonjs": "^17.1.0",
     "@rollup/plugin-node-resolve": "^11.2.0",
+    "@rollup/plugin-typescript": "^8.2.0",
+    "@types/bytes": "^3.1.0",
+    "@types/d3-array": "^2.9.0",
+    "@types/d3-color": "^2.0.1",
+    "@types/d3-hierarchy": "^2.0.0",
+    "@types/d3-scale": "^3.2.2",
+    "@types/d3-shape": "^2.0.0",
+    "@types/nanoid": "^2.1.0",
+    "@typescript-eslint/eslint-plugin": "^4.19.0",
+    "@typescript-eslint/parser": "^4.19.0",
     "bytes": "^3.1.0",
     "d3-array": "^2.12.0",
     "d3-color": "^2.0.0",
@@ -45,6 +54,9 @@
     "d3-shape": "^2.1.0",
     "del-cli": "^3.0.1",
     "eslint": "^7.22.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-react": "^7.23.1",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "htm": "^3.0.4",
     "postcss": "^8.2.8",
     "postcss-url": "^10.1.3",
@@ -54,9 +66,11 @@
     "rollup-plugin-postcss": "^4.0.0",
     "rollup-plugin-terser": "^7.0.2",
     "sass": "^1.32.8",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.2.3",
     "webcola": "^3.4.0"
   },
   "engines": {
-    "node": ">=10"
+    "node": ">=10.16"
   }
 }
