@@ -1,17 +1,22 @@
-"use strict";
+import { promises as fs } from "fs";
+import path from "path";
+import { VisualizerData } from "../types/types";
+import { TemplateType } from "./template-types";
 
-const fs = require("fs").promises;
-const path = require("path");
-
-const htmlEscape = (string) =>
-  string
+const htmlEscape = (str: string) =>
+  str
     .replace(/&/g, "&amp;")
     .replace(/"/g, "&quot;")
     .replace(/'/g, "&#39;")
     .replace(/</g, "&lt;")
     .replace(/>/g, "&gt;");
 
-const buildHtmlTemplate = ({ title, script, nodesData, style }) =>
+const buildHtmlTemplate = (
+  title: string,
+  script: string,
+  nodesData: string,
+  style: string
+) =>
   `
 <!DOCTYPE html>
 <html lang="en">
@@ -40,7 +45,7 @@ ${script}
       const height = window.innerHeight;
 
       const chartNode = document.querySelector("main");
-      drawChart(chartNode, data, width, height);
+      drawChart.default(chartNode, data, width, height);
     };
 
     window.addEventListener('resize', run);
@@ -53,16 +58,21 @@ ${script}
 
 `;
 
-module.exports = async function buildHtml({ title, data, template }) {
+interface BuildHtmlOptions {
+  title: string;
+  data: VisualizerData;
+  template: TemplateType;
+}
+
+export async function buildHtml({
+  title,
+  data,
+  template,
+}: BuildHtmlOptions): Promise<string> {
   const [script, style] = await Promise.all([
     fs.readFile(path.join(__dirname, "..", "lib", `${template}.js`), "utf8"),
     fs.readFile(path.join(__dirname, "..", "lib", `${template}.css`), "utf8"),
   ]);
 
-  return buildHtmlTemplate({
-    title,
-    style,
-    script,
-    nodesData: JSON.stringify(data),
-  });
-};
+  return buildHtmlTemplate(title, script, JSON.stringify(data), style);
+}
