@@ -1,11 +1,15 @@
-export const LABELS = {
+import { SizeKey, VisualizerData } from "../types/types";
+
+export const LABELS: Record<SizeKey, string> = {
   renderedLength: "Rendered",
   gzipLength: "Gzip",
   brotliLength: "Brotli",
 };
 
-export const getAvailableSizeOptions = (options = {}) => {
-  const availableSizeProperties = ["renderedLength"];
+export const getAvailableSizeOptions = (
+  options: VisualizerData["options"]
+): SizeKey[] => {
+  const availableSizeProperties: SizeKey[] = ["renderedLength"];
   if (options.gzip) {
     availableSizeProperties.push("gzipLength");
   }
